pyproject.toml
[project]
name = "bitharbor"
version = "0.1.0"
description = "BitHarbor local-first media server backend"
authors = [
  { name = "BitHarbor Team", email = "team@bitharbor.local" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "sqlalchemy>=2.0.30",
  "sqlalchemy-utils>=0.41.2",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
  "aiosqlite>=0.20.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt[crypto]>=2.8.0",
  "python-multipart>=0.0.9",
  "blake3>=0.4.1",
  "numpy>=1.26.4",
  "scipy>=1.13.0",
  "pillow>=10.3.0",
  "opencv-python-headless>=4.9.0.80",
  "ffmpeg-python>=0.2.0",
  "imageio>=2.34.0",
  "librosa>=0.10.2",
  "torchaudio>=2.2.2",
  "torchvision>=0.17.2",
  "torch>=2.2.2",
  "imagebind>=0.2.1",
  "open-clip-torch>=2.23.0",
  "diskannpy>=0.9.0",
  "yt-dlp>=2024.5.27",
  "httpx>=0.27.0",
  "rich>=13.7.1",
  "orjson>=3.10.0",
  "python-dateutil>=2.9.0",
  "typer>=0.12.3",
  "psutil>=5.9.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "pytest-cov>=5.0.0",
  "httpx[http2]>=0.27.0",
  "ruff>=0.4.4",
  "mypy>=1.10.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E",
  "F",
  "W",
  "I001",
  "C90"
]
ignore = [
  "E501"
]

[tool.ruff.isort]
known-first-party = ["app", "api", "db", "domain", "features", "infrastructure", "utils"]
profile = "black"

